==> 01_Modeling_Parameters/03_Gas_stations/01_new_service_stations.R <==
library(tidyverse)
library(RODBC)

odbcDataSources()

credentials <- read.csv("X:/Agency_Files/Outcomes/Risk_Eval_Air_Mod/_Air_Risk_Evaluation/R/R_Camp/Student Folder/credentials.csv", stringsAsFactors = FALSE)

d_cnx <- odbcConnect("deltaw", uid = credentials$delta_user, pwd = credentials$delta_pwd, believeNRows = FALSE)

odbcGetInfo(d_cnx)

service_stations  <- sqlQuery(d_cnx, "SELECT WH_TEMPO.MV_SI_METADATA_MF.MASTER_AI_NAME,
  WH_TEMPO.MV_SI_METADATA_MF.MASTER_AI_ID,
  WH_TEMPO.MV_SI_METADATA_MF.ADDRESS1,
  WH_TEMPO.MV_SI_METADATA_MF.CITY_NAME,
  WH_TEMPO.MV_SI_METADATA_MF.STATE_CODE,
  WH_TEMPO.MV_SI_METADATA_MF.ZIP_CODE,
  WH_TEMPO.MV_SI_METADATA_MF.COUNTY_NAME,
  TEMPO.MTB_SUBSECTOR.SUBSECTOR_DESC,
  WH_TEMPO.MV_SI_METADATA_MF.STATUS,
  WH_TEMPO.MV_SI_METADATA_MF.LATITUDE,
  WH_TEMPO.MV_SI_METADATA_MF.LONGITUDE,
  TEMPO.AGENCY_INTEREST_ALT.ALTERNATE_AI_ID,
  TEMPO.MTB_USER_GROUP.USER_GROUP_DESCRIPTION,
  APP_INTERNAL.SPATIAL_SI_MF.SPATIAL_ID,
  APP_INTERNAL.SPATIAL_SI_MF.METHOD_DESC,
  WH_TEMPO.MV_SI_METADATA_MF.TMSP_CREATED,
  WH_TEMPO.MV_SI_METADATA_MF.TMSP_LAST_UPDT,
  WH_TEMPO.MV_SI_METADATA_MF.USER_LAST_UPDT,
  TEMPO.AGENCY_INTEREST_ALT.START_DATE
FROM WH_TEMPO.MV_SI_METADATA_MF
INNER JOIN TEMPO.SUBJ_ITEM_SECTOR
ON WH_TEMPO.MV_SI_METADATA_MF.MASTER_AI_ID                = TEMPO.SUBJ_ITEM_SECTOR.MASTER_AI_ID
AND WH_TEMPO.MV_SI_METADATA_MF.INT_DOC_ID                 = TEMPO.SUBJ_ITEM_SECTOR.INT_DOC_ID
AND WH_TEMPO.MV_SI_METADATA_MF.SUBJECT_ITEM_CATEGORY_CODE = TEMPO.SUBJ_ITEM_SECTOR.SUBJECT_ITEM_CATEGORY_CODE
AND WH_TEMPO.MV_SI_METADATA_MF.SUBJECT_ITEM_ID            = TEMPO.SUBJ_ITEM_SECTOR.SUBJECT_ITEM_ID
INNER JOIN TEMPO.MTB_SUBSECTOR
ON TEMPO.MTB_SUBSECTOR.SUBSECTOR_CODE = TEMPO.SUBJ_ITEM_SECTOR.SUBSECTOR_CODE
INNER JOIN TEMPO.MTB_SECTOR
ON TEMPO.MTB_SECTOR.SECTOR_CODE = TEMPO.SUBJ_ITEM_SECTOR.SECTOR_CODE
INNER JOIN TEMPO.AGENCY_INTEREST_ALT
ON WH_TEMPO.MV_SI_METADATA_MF.MASTER_AI_ID = TEMPO.AGENCY_INTEREST_ALT.MASTER_AI_ID
AND WH_TEMPO.MV_SI_METADATA_MF.INT_DOC_ID  = TEMPO.AGENCY_INTEREST_ALT.INT_DOC_ID
INNER JOIN TEMPO.MTB_USER_GROUP
ON TEMPO.MTB_USER_GROUP.USER_GROUP_ID = TEMPO.AGENCY_INTEREST_ALT.USER_GROUP_ID
INNER JOIN APP_INTERNAL.SPATIAL_SI_MF
ON APP_INTERNAL.SPATIAL_SI_MF.ITEM_ID        = WH_TEMPO.MV_SI_METADATA_MF.ITEM_ID
WHERE TEMPO.MTB_SUBSECTOR.SUBSECTOR_DESC     = 'Service station'
AND WH_TEMPO.MV_SI_METADATA_MF.STATUS        = 'Active'
AND (TEMPO.AGENCY_INTEREST_ALT.USER_GROUP_ID = 'UT+'
OR TEMPO.AGENCY_INTEREST_ALT.USER_GROUP_ID   = 'AT+')
AND TEMPO.AGENCY_INTEREST_ALT.END_DATE      IS NULL
ORDER BY WH_TEMPO.MV_SI_METADATA_MF.MASTER_AI_NAME
", stringsAsFactors = FALSE)

write_csv(service_stations, "X:/Agency_Files/Outcomes/Risk_Eval_Air_Mod/_Air_Risk_Evaluation/MNRISKS/2017 MNRISKS/_development/1. Modeling parameters/3. Gas stations/service_station_tempo_query.csv")
